+++ Runtime/UI/Components/ColorField.cs
@@ -56,6 +56,8 @@ namespace Unity.AppUI.UI
 
         internal static readonly BindingId hdrProperty = nameof(hdr);
 
+        internal static readonly BindingId clickableProperty = nameof(clickable);
+
 #endif
 
         /// <summary>
@@ -294,12 +296,17 @@ namespace Unity.AppUI.UI
             get => m_Clickable;
             set
             {
+                var changed = m_Clickable != value;
                 if (m_Clickable != null && m_Clickable.target == this)
                     this.RemoveManipulator(m_Clickable);
                 m_Clickable = value;
                 if (m_Clickable == null)
                     return;
                 this.AddManipulator(m_Clickable);
+#if ENABLE_RUNTIME_DATA_BINDINGS
+                if (changed)
+                    NotifyPropertyChanged(in clickableProperty);
+#endif
             }
         }