+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0-pre.22] - 2025-06-17
+
+### Added
+
+- Added support of `ICommand` in `Pressable` manipulator.
+- Added runtime binding support for commands in `Pressable` via `clickable.command` property in most interactable UI elements.
+
 ## [2.0.0-pre.21] - 2025-06-13
 
 ### Changed